feat : 로그인 성공하면 home 화면 (issue 목록 화면) 으로 이동하도록 구현
- UI 레이어에 맞게 디렉토리 생성 및 리팩터링
