+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.issu_tracker">
 
-    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -12,19 +12,19 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Issutracker">
         <activity
-            android:name=".HomeActivity"
-            android:exported="true" >
+            android:name=".ui.home.HomeActivity"
+            android:exported="true">
+
+        </activity>
+
+        <activity
+            android:name=".ui.login.LogInActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-
-        <activity
-            android:name=".MainActivity"
-            android:exported="true">
-
         </activity>
     </application>
