+++ app/src/main/java/com/example/issu_tracker/ui/issue/IssueFragment.kt
@@ -1,17 +1,19 @@
-package com.example.issu_tracker
+package com.example.issu_tracker.ui.issue
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
+import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.issu_tracker.R
 import com.example.issu_tracker.databinding.FragmentIssueBinding
+import com.example.issu_tracker.ui.home.HomeViewModel
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
