+++ app/src/main/res/layout/activity_home.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".HomeActivity">
+        tools:context=".ui.home.HomeActivity">
 
         <fragment
             android:id="@+id/nav_host"
