--- app/src/main/res/layout/activity_main.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/btn_google_sign_in"
-        style="@style/Body2.white"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="@color/black"
-        android:drawableStart="@drawable/ic_google_logo"
-        android:drawablePadding="8dp"
-        android:paddingHorizontal="12dp"
-        android:text="@string/google_sign_in"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/btn_github_sign_in"
-        style="@style/Body2.white"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:background="@color/black"
-        android:drawableStart="@drawable/ic_github_logo"
-        android:drawablePadding="8dp"
-        android:paddingHorizontal="12dp"
-        android:text="@string/github_sign_in"
-        app:layout_constraintLeft_toLeftOf="@+id/btn_google_sign_in"
-        app:layout_constraintRight_toRightOf="@+id/btn_google_sign_in"
-        app:layout_constraintTop_toBottomOf="@+id/btn_google_sign_in" />
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/btn_sign_out"
-        style="@style/Body2.white"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:background="@color/black"
-        android:paddingHorizontal="12dp"
-        android:text="로그아웃"
-        app:layout_constraintLeft_toLeftOf="@+id/btn_github_sign_in"
-        app:layout_constraintRight_toRightOf="@+id/btn_github_sign_in"
-        app:layout_constraintTop_toBottomOf="@+id/btn_github_sign_in" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
