+++ app/src/main/res/layout/fragment_issue.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:padding="19dp"
-        tools:context=".IssueFragment">
+        tools:context=".ui.issue.IssueFragment">
 
         <ImageButton
             android:id="@+id/iv_issue_menu"
