+++ app/src/main/res/layout/item_issue.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="issue"
-            type="com.example.issu_tracker.Issue" />
+            type="com.example.issu_tracker.data.Issue" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
